Change type of uspace pointers in kernel from pointer type to numeric (#170)
From kernel's perspective, userspace addresses are not valid pointers,
and can only be used in calls to `copy_to/from_uspace()`.
Therefore, we change the type of those arguments and variables to
`uspace_addr_t` which is an alias for `sysarg_t`.

This allows the compiler to catch accidental direct accesses to
userspace addresses.

Additionally, to avoid losing the type information in code,
a macro `uspace_ptr(type)` is used that translates to `uspace_addr_t`.
I makes no functional difference, but allows keeping the type information
in code in case we implement some sort of static checking for it in the future.

However, ccheck doesn't like that, so instead of using `uspace_ptr(char)`,
we use `uspace_ptr_char` which is defined as
`#define uspace_ptr_char uspace_ptr(char)`.

+++ kernel/arch/arm64/src/arm64.c
@@ -155,11 +155,11 @@ void userspace(uspace_arg_t *kernel_uarg)
 	    SPSR_MODE_ARM64_EL0T);
 
 	/* Set program entry. */
-	ELR_EL1_write((uint64_t) kernel_uarg->uspace_entry);
+	ELR_EL1_write(kernel_uarg->uspace_entry);
 
 	/* Set user stack. */
-	SP_EL0_write(((uint64_t) kernel_uarg->uspace_stack +
-	    kernel_uarg->uspace_stack_size));
+	SP_EL0_write(kernel_uarg->uspace_stack +
+	    kernel_uarg->uspace_stack_size);
 
 	/* Clear Thread ID register. */
 	TPIDR_EL0_write(0);
